Refactor: Use Int64 IDs and domain names across all layers
Standardize all entity IDs to Int64 (long) and replace legacy property names (e.g., RowId, RowDesc) with clear, domain-specific names (e.g., BranchId, BranchNameAr, RoleNameEn) in DTOs, commands, queries, repositories, controllers, and tests. Update validation and Oracle data access accordingly. No business logic is changed; this improves type safety, clarity, and consistency throughout the codebase. Test code is updated to match the new conventions.

+++ src/ThinkOnErp.Application/Features/Users/Commands/ChangePassword/ChangePasswordCommand.cs
@@ -4,7 +4,7 @@ namespace ThinkOnErp.Application.Features.Users.Commands.ChangePassword;
 
 public class ChangePasswordCommand : IRequest<bool>
 {
-    public decimal UserId { get; set; }
+    public Int64 UserId { get; set; }
     public string CurrentPassword { get; set; } = string.Empty;
     public string NewPassword { get; set; } = string.Empty;
     public string ConfirmPassword { get; set; } = string.Empty;
